217. React Cooperating with Redux(Section 17: Integrating React with Redux)
All right my friends I think that we've now got a much better idea on how redux works on its own by

writing some code into a little application on Copen it's never going to start putting together a re-act

app and then eventually integrate redux into it.

This is going to give you a much better idea on how the two libraries interact with each other and how

we can use them to form a fully fledged complex application.

Now I'll be honest with you at the first hour we're going to put together is very simple very straightforward.

We're going to put together this very simple out very quickly so that we can then move on to much more

complicated stuff after we've gotten a little bit of a practice on the easy application we're about

to put together.

So let's take a look at the market.

Here it is.

So on the left hand side this application we're going to have a hard coded list of songs.

So we're just going to list out the title of a couple of different songs right here.

Well then show a select button next to each one whenever a user clicks on that select button we're then

going to show details for that song on the right hand side of the screen.

If the user then clicks on another select button will show details for that song.

So like I said simple straightforward just to help us really understand how to put react in redux together.

So without further ado let's flip on over to our terminal and we're going to generate a new project

selling enough foot over.

You'll notice that I'm now back inside my workspace directory so I'm no longer inside that videos project

folder inside of here or generate a new app with create react at I'll call this one simply.

Songs like so.

All right.

That's going to start the install process.

So let's take a quick pause right here.

When we come back the next section we're going to talk about how we're going to approach the design

of this application from the component aspect.

So quick break and I'll see you in just a minute.