218.1 React, Redux, and...React-Redux! {yarn add redux react-redux --save}(Section 17: Integrating React with Redux)
In the last section we generated a new project now we called Songs.

So not going to change into that project directory like so now I'm not going to start up my project

here just yet because I want to show you one or two quick things that we're going to have to do before

we actually start the project up.

All right.

So back over here a couple of quick diagrams.

First I want to show you how we're going to break out this application into separate components.

So we're going to make one component for the right hand side.

We're going to call it a song list.

It's going to take a list of songs as you might guess those will be an array of objects for each object

has maybe the title of the song and maybe the length as well.

We will then render out a list of all those different songs and then for the right hand side we'll have

a very simple song detail if it is given a song object it will display the title and the length of the

given song.

And that's pretty much it.

So without a doubt the Riak side of this application.

Simple straightforward.

All right.

Now when I tell you something a little bit more interesting that's going to help you understand how

we put react in redux together.

So at this point we've made use of the re-act library by itself quite a bit.

We also have seen an example of redux inside of Copan and you even saw me and you probably went through

the same process as well at redux to that code instance as a dependency.

Now we're going to flip back over toward terminal's here and just a second and we're going to install

redux but at the same time we're going to install a second library called re-act redux.

That's literally it's name it's re-act dash redux.

This is a third library that we're going to use to make sure that react in redux can talk to each other.

Remember redux was not necessarily designed to work only with react so the re-act redux library has

a bunch of helper functions inside of it to get redux to work nicely with you guessed it react we're

going to really be learning a lot about re-act redux at this point and moving forward because we already

know how re-act works pretty well and he's got a reasonable idea of exactly what redux is doing.

So in this app what we're really learning at the end of the day is how re-act redux works as opposed

to just react or redux.

OK so we're going to flip back over toward terminal.

I'm going to make sure I'm inside that Sung's project directory and I'll do an NPM install dash dash

save redux and re-act dash redux like so.

All right now this is just going to take a second or two to install so I'm just going to let it do its

thing.

I know that's so repetitive but let's take a break and we're going to figure out how that thing works

in the next section.