218.2 React, Redux, and...React-Redux! (Section 17: Integrating React with Redux)
And then when it's done I'm going to start up NPM start inside this project directory.

There we go.

It's all run NPM start and it will see our browser open up and here we go localhost 3000 like so.

Next up I'm going to start up my code editor inside of our project directory.

And as you are very well used to doing.

We're going to find that SIRC folder and we're going to delete everything inside there.

And then inside the US RC folder will go back in and we're going to create our root index js file.

I'm going to put together a little bit of the plate inside of your soul say import re-act from react

or do import re-act Dom from re-act dash Dom and then we don't have a app component or a app file yet

but we're going to make it in just a second so I'm going to imagine that already exists and I'll say

import app from components slash app like so and then at the bottom we'll do the classic re-act render

with our app and document query selector with root like so.

All right.

So remember at this point we don't have our app component put together so if you flip back over to the

browser you're probably going to see an error message.

So let's now very quickly create our app component.

So going to find my or directory make a new folder inside they're called components and inside of that

folder I'm going to make a new file called.

J.S..

And then inside of here we will import re-act from react.

We'll create a functional component this time old app and inside of it or turn a simple div with a text

app and then export default app at the bottom like so that's it.

That's all we do.

So let's do a quick check inside the browser and I should now see the text app appear on the screen.

Perfect.

Now one last little bit of setup that I want to do as long as we are going through all this kind of

repetitive stuff.

I want to make sure that I hook up semantic UI just the CSSA file to our project as well.

So I'm sure you remember how to do this.

We're going to do a quick search for semantic UI CDN.

I'll find the Cdn.

Dot com link.

And then I'll do a search for semantic men CSA.

Siri go right there.

So I'm going to copy that thing.

I will flip flip back over to my public index.

HS e-mail file.

And then just as before I'll put in my link with the rel style sheets and my eight's ref to the link

that I just copied.

Like so and I'll save the file and I'll just make sure that I see the updated font for my app text that

appears appear on the screen.

All right that's it.

So we're now ready to start putting together our actual app and learning how to get react in redux to

work together with that re-act dash redux library.