226.1. Configuring Connect with MapStateToProps(Section 17: Integrating React with Redux)
In less sexually defined a connect function inside of our song list component.

So remember this connecting right here even though it doesn't really look like it is actually a re-act

component.

It's a reaction upon it that we're going to pass some configuration to.

We're going to specifically tell this connecting that we want to get a list of songs out of our redux

store from the provider.

So anytime that our list of songs and side of our store changes that provider is going to automatically

notify our connect function right here and the connect function is then going to pass our list of songs

down to our song list component.

So let's now figure out how we can configure this connect function or connect component and tell it

that we want to get some specific data out of our redux store.

All right so to do so we're going to come back over here and I'm going to define a function right above

our export default statement.

Now by convention this is Convention we normally call this function the map state to prop's function

so now this can be an arrow function.

It could be a function keyword.

You will see it defined both ways map state to prop's.

So this is a very crazy little function name.

But if you really kind of read it it will eventually start to make a little bit more sense.

This is saying that we're going to take our state object essentially all of the data that's inside of

our redux store and we're going to run some calculation on it or some computation that's going to cause

that data to eventually show up as props inside of our component.

So that's the meaning of map state to process this function does not have to be called Map state to

prop's we could just as easily call it get my states like so we can call it anything we want but by

convention you're usually going to see this defined as something called Map state to prop's this map

state to prop's function is going to be called with all of the state inside of a redux store.

So that state right there is essentially going to be our entire list of songs from the song list reducer

and are currently selected songs from the selected song reduce her state right here.

All of that data inside of a redux store if we were still working on that insurance application the

state object right here would contain our accounts and our policies and our claims history properties

as well.

So then right now let's try just cons. logging that thing out.

I mean to say console log States and then after that I'm going to say return state just to make sure

that everything works just fine.

Now after defining that function and then going to take it and pass it as the first argument to this

connector function down here.

So that's how we can figure this connect component.

I've said several times oh yeah we need to configure connect in some way and tell it how to get some

data from provider.

We can figure it by passing it a function it's going to pass that as the first function to connect that

state to props.

Now again I want to remind you we can call this anything we want to call it so we could call it get

my state if we want to.

Nothing special about calling it map state to prop's.

All right so let's save this.

We're going to run it and see what happens.

All right so I can flip back over and you'll immediately notice that we've got our state object printed

out right here.

We've got selected Saag that has all the data that was returned from our selected song reducer which

by default was no.

And then we've got these songs Saray as well that has those four song objects that we had to.

So now we essentially want to pull off just this song's property and make sure that it somehow gets

into our song list component.