226.2. Configuring Connect with MapStateToProps(Section 17: Integrating React with Redux)
Remember the song component doesn't care about the selected song at all it just cares about songs.

So to do so inside of my mouth state to prop's function right here I'm going to remove that council

Lague and I'm going to remove the state right there and I'm going to replace it with an object this

object is going to have a key of songs and its value will be state DOT songs like so so the object that

we return from the map state to prop's function is going to show up as props inside of our component

three essentially Sayne by returning the subject right here where essentially saying that at some point

in time the future inside of our songless component if we say this start prop's it's going to be equal

to exactly that object.

Right.

They're going to be an object that has a song's property and that will contain the list or the array

of all of our songs.

So that is exactly how we make use of the re-act redux library to get data from our redux store into

a component.

It's always going to be the exact same scenario we're always going to import connect at the top.

We're always going to call connect and then pass in our component as the second function call.

We're always going to define map state to prop's map state to prop's is always going to get a first

argument of state and we're always going to return an object that is going to show up as props inside

of our component.

Now if any of that seemed mysterious then just do a quick replay of those last 15 seconds or if you

missed anything in there to replay those last 15 seconds because that entire flow that I just mentioned

is going to be repeated a billion times.

On every single re-act project they you ever work on.

It's always going to be the same code the see inside of your.

Over and over and over again always import connects always the component always maps state to props

and then always export default connect your components right there.

All right so with that further ado inside of my components I'm going to try doing a console log of this

props like so let's just see what data shows up inside that component.

So I'll flip back over.

And now here's my prop's object.

I see my list of songs.

There is the list of songs right there.

And then you'll notice that we also got a reference to the dispatch function from our redux store as

well.

This is the same dispatch function that we use previously inside of code pen to manually dispatch an

action creator or someone manually dispatched an action after calling an action creator.

So if you ever want to change some data inside of our application we can make use of this dispatch function

right here.

But as you'll see in just a little bit there's a much more convenient way of doing that.

All right.

Now that we've got our list of songs inside of here please take a pause.

When we come back in the next section we're going to render this list of songs out inside of our song

list component.

So quick break and I'll see you in just a minute.